requirements.txt
python-docx==0.8.11
# googletrans==4.0.0-rc1
openai==0.27.0
